Move DialogQueue into common/
(cherry picked from commit 0ee3e00ed4ee1c13c9268dedc45b58a7c2abcfe5)
